BLN: Make it blink!
-The way BLN is now seem pretty useless. It just turns on
and waits for the typical led time out (Normally 5 seconds). If
you weren't looking at it, you'd miss it.

-Now, if you recieve a notification while the screen is off. BLN
blinks until you turn the screen on.

-Further considerations: There's room in here to have custom blink
patterns, so maybe in the future... Perhaps there should be code
in here to check battery level, i.e. no blink if battery is too low...
Perhaps no blink if "in pocket" (proximity) detected...